requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# ML Core (compatible versions)
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Deep Learning
torch>=2.0.0
transformers>=4.30.0
tensorflow>=2.13.0

# NewsAPI
newsapi-python==0.2.7
requests>=2.31.0
python-dotenv>=1.0.0

# NLP & Similarity
sentence-transformers>=2.2.0

# Visualization
plotly>=5.17.0